+++ src/libxpeccy/hardware/plus2.c
@@ -37,10 +37,6 @@ void plusRes(Computer* comp) {
 
 // in
 
-int p2InFF(Computer* comp, int port) {
-	return (comp->vid->vbrd || comp->vid->hbrd) ? 0xff : comp->vid->atrbyte & 0xff;
-}
-
 // no upd765 in +2
 int p2_dos_rd(Computer* comp, int port) {
 	return -1;
@@ -76,7 +72,7 @@ static xPort p2PortMap[] = {
 	{0xe002,0x2ffd,2,2,2,p2_dos_rd,	p2_dos_wr},	// 2ffd/3ffd
 	{0xc002,0xbffd,2,2,2,NULL,	xOutBFFD},
 	{0xc002,0xfffd,2,2,2,xInFFFD,	xOutFFFD},
-	{0x0000,0x0000,2,2,2,p2InFF,	NULL}
+	{0x0000,0x0000,2,2,2,zx_in_float,	NULL}
 };
 
 void pl2Out(Computer* comp, int port, int val) {
@@ -98,4 +94,4 @@ xPortDsc pl2_port_tab[] = {
 };
 
 HardWare pl2_hw_core = {HW_PLUS2A,"Plus2A","Spectrum +2A",MEM_128K,1.0,NULL,pl2_port_tab,
-			zx_init,pl2MapMem,pl2Out,pl2In,stdMRd,stdMWr,zx_irq,zx_ack,plusRes,zx_sync,zx_keyp,zx_keyr,zx_vol};
+			zx_init,pl2MapMem,pl2Out,pl2In,asicMRd,asicMWr,zx_irq,zx_ack,plusRes,zx_sync,zx_keyp,zx_keyr,zx_vol};